Fix locale for `file-name-from-day`
The following tests were failing because the function `file-name-from-day` insisted in adding a dot (`.`) to the month:

 - data-storage partition-messages-by-day calculates file name from day
 - The Migrator migration 7 break message-file into daily files breaks up message-file and deletes it

I was running the tests in my machine with local set to Canada.

Considering that it is for an automated file name creation, maybe full consideration of locale is not necessary. So, fixing it to the expected output by the Original Author makes sense.