Get open-link to work (when user clicks on a URL).
It didn't do anything except print an error message when I tried clicking
on links in the article window. I assume this is because browse/browse-url
expects a string and .getURL returns some kind of url object, and the two
are not compatible.

The error message was:
  Exception in thread "AWT-EventQueue-0" java.lang.ClassCastException:
  class java.net.URL cannot be cast to class java.lang.String (java.net.URL and java.lang.String are in module java.base of loader 'bootstrap')
	at clojure.java.browse$open_url_in_browser.invokeStatic(browse.clj:50)
	at clojure.java.browse$browse_url.invokeStatic(browse.clj:74)
	at clojure.java.browse$browse_url.invoke(browse.clj:66)

(This occurred with Clojure 1.10.3, Java 18.0.1.1., Windows 10)